Make Requirements less misleading
MinGW is a port of GCC (and related tools) to Windows.
MSYS2 is a port of a bunch of Unix tools to Windows, and in a full install, includes MinGW.
Git for Windows includes a stripped-down MSYS2 install, which incudes Bash, but doesn't include MinGW's C++ compiler.

This project requires a C++ compiler (e.g. MinGW) and a Bash shell (e.g. MSYS2's, which comes with Git for Windows).

Buddy: Sam